Split up tests; filtered anonymous accounts out of user lists
